Feature gate for shared
