+++ broker/Cargo.toml
@@ -44,7 +44,7 @@ once_cell = "1.17.1"
 bytes = { version = "1.4.0", optional = true }
 
 [features]
-sockets = ["dep:bytes"]
+sockets = ["dep:bytes", "shared/sockets"]
 
 [build-dependencies]
 build-data = "0"
