+++ proxy/Cargo.toml
@@ -47,7 +47,7 @@ chacha20poly1305 = { version = "0.10.1", features = ["stream"], optional = true 
 bytes = { version = "1.4", optional = true}
 
 [features]
-sockets = ["dep:bytes", "dep:chacha20poly1305", "tokio-util/codec", "tokio-util/compat"]
+sockets = ["dep:bytes", "dep:chacha20poly1305", "tokio-util/codec", "tokio-util/compat", "shared/sockets"]
 
 [build-dependencies]
 build-data = "0"
