+++ shared/Cargo.toml
@@ -60,6 +60,7 @@ fundu = "0.5.0"
 tokio-test = "0.4.2"
 
 [features]
+sockets = []
 default = []
 config-for-proxy = []
 config-for-central = []
