+++ tests/Cargo.toml
@@ -7,7 +7,7 @@ license = "Apache-2.0"
 
 [dependencies]
 tokio = "1.28.0"
-shared = {path = "../shared"}
+shared = {path = "../shared", features = ["sockets"]}
 once_cell = "1.17.1"
 serde_json = "1.0.96"
 http = "0.2.9"
